refactor(core): tidy up the throttling manager's public surface
- `parseRetryAfterHeader` was public API exported from a storages module despite
  being a plain HTTP header parser, so it moves to `http.ts` alongside its test.
- Adds `innerManager`, so a caller that hands the wrapper its queue can still get
  the queue back, matching crawlee-python's `inner`.
- `domains` now rejects empty strings instead of quietly dropping them, which
  turned a typo into a silently unthrottled domain.
- Makes the two remaining time-sensitive tests deterministic: the backoff decay
  test slept out 20ms windows and flaked under load, and the purge test never
  checked the sub-queue was non-empty to begin with.
